iceberg-geo-api/pyproject.toml
[project]
name = "iceberg-geo-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyiceberg[pyarrow,duckdb,s3]>=0.11.0",
    "duckdb>=1.1.0",
    "shapely>=2.0",
    "pyproj>=3.6",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pygeoapi = [
    "pygeoapi>=0.22.0",
    "geoarrow-pyarrow>=0.1.0",
]
geoservices = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "protobuf>=4.25",
    "python-multipart>=0.0.9",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pyarrow>=15.0",
    "geopandas>=0.14",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
